Extract some types to their own files
